Run Angular Project With API

+++ Shared/Basket/BasketItemDTO.cs
@@ -12,9 +12,9 @@ namespace Shared.Basket
         public int Id { get; init; }
         public string ProductName { get; init; }
         public string PictureUrl { get; init; }
-        [Range(1,byte.MaxValue)]
+        [Range(1,short.MaxValue)]
         public decimal Price { get; init; }
-        [Range(1, byte.MaxValue)]
+        [Range(1, short.MaxValue)]
         public int Quantity { get; init; }
     }
 }
